Bug 1404251 - Fix various UBSan issues in SpiderMonkey. r=bbouvier

+++ js/src/jit/IonAnalysis.cpp
@@ -3207,7 +3207,7 @@ jit::ExtractLinearSum(MDefinition* ins, MathSpace space)
     if (ins->isAdd()) {
         int32_t constant;
         if (space == MathSpace::Modulo)
-            constant = lsum.constant + rsum.constant;
+            constant = uint32_t(lsum.constant) + uint32_t(rsum.constant);
         else if (!SafeAdd(lsum.constant, rsum.constant, &constant))
             return SimpleLinearSum(ins, 0);
         return SimpleLinearSum(lsum.term ? lsum.term : rsum.term, constant);
@@ -3218,7 +3218,7 @@ jit::ExtractLinearSum(MDefinition* ins, MathSpace space)
     if (lsum.term) {
         int32_t constant;
         if (space == MathSpace::Modulo)
-            constant = lsum.constant - rsum.constant;
+            constant = uint32_t(lsum.constant) - uint32_t(rsum.constant);
         else if (!SafeSub(lsum.constant, rsum.constant, &constant))
             return SimpleLinearSum(ins, 0);
         return SimpleLinearSum(lsum.term, constant);
